refactor: update jwt middleware import path
