+++ middleware/jwt/errors.go
@@ -1,4 +1,4 @@
-package expressgo_jwt
+package jwt
 
 import (
 	"net/http"
